Create separate reader file
