+++ main/src/Poll.java
@@ -1,11 +1,5 @@
-import java.io.BufferedReader;
-import java.io.FileReader;
-import java.io.IOException;
 import java.time.LocalDate;
-import java.time.format.DateTimeFormatter;
 import java.time.temporal.ChronoUnit;
-import java.util.ArrayList;
-import java.util.List;
 
 public class Poll {
 
@@ -38,30 +32,6 @@ public class Poll {
         this.daysBeforeElection = ChronoUnit.DAYS.between(dateTaken, ELECTION_DATE);
         this.pollsterName = pollsterName;
         this.standardDeviation = Math.sqrt(demPercent * (1 - demPercent) / sampleSize);
-    }
-
-    public static Poll[] readNationalPolls(String filename) throws IOException {
-        String line;
-        List<Poll> polls = new ArrayList<>();
-        BufferedReader reader = new BufferedReader(new FileReader(filename));
-        //Clear header line
-        reader.readLine();
-        while ((line = reader.readLine()) != null) {
-            String[] commaSplit = line.split(",");
-            LocalDate date = LocalDate.parse(commaSplit[0], DateTimeFormatter.ofPattern("M/d/yyyy"));
-            double rawDemPercent = Double.parseDouble(commaSplit[1]);
-            double rawRepPercent = Double.parseDouble(commaSplit[2]);
-            double sampleSize = Double.parseDouble(commaSplit[3]);
-            boolean registeredVoter = Boolean.parseBoolean(commaSplit[4]);
-            double houseLean = Double.parseDouble(commaSplit[5]);
-            Grade grade = Grade.parseGrade(commaSplit[6]);
-            String pollsterName = commaSplit[7];
-            polls.add(new Poll(date, rawDemPercent, rawRepPercent, sampleSize, registeredVoter, houseLean, grade,
-                    pollsterName));
-        }
-        reader.close();
-
-        return polls.toArray(new Poll[1]);
     }
 
     public LocalDate getDateTaken() {
